src/python/google_drive/requirements.txt
google-auth>=2.20.0
google-auth-oauthlib>=0.5.2
google-auth-httplib2>=0.1.0
google-api-python-client>=2.78.0
